dom/messagechannel/tests/test_messageChannel_bug1224825.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1224825
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1224825</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1224825">Mozilla Bug 1224825</a>
<div id="content"></div>
<pre id="test">
</pre>
  <script type="application/javascript">

  var MAX = 100;

  function runTest() {
    var worker = new Worker('data:javascript,onmessage = function(e) { e.ports[0].onmessage = function(evt) { postMessage(evt.data);}}');

    var count = 0;
    worker.onmessage = function(e) {
      is(e.data, count, "Correct value expected!");
      ok(count < MAX,"No count > MAX messages!");
      if (++count == MAX) {

        SimpleTest.requestFlakyTimeout("Testing an event not happening");
        setTimeout(function() {
          SimpleTest.finish();
        }, 200);

        info("All the messages correctly received");
      }
    }

    var mc = new MessageChannel();
    worker.postMessage(42, [mc.port2]);

    for (var i = 0; i < MAX; ++i) {
      mc.port1.postMessage(i);
    }

    mc.port1.close();

    for (var i = 0; i < MAX * 2; ++i) {
      mc.port1.postMessage(i);
    }

  }

  SimpleTest.waitForExplicitFinish();
  runTest();
  </script>
</body>
</html>
